semi-broken implementation of selection and dragging
from [I like to move it](https://web.archive.org/web/20170508215902/http://www.trackze.ro/wsbe-3-i-like-to-move-it/).

got selection and dragging working, but it's half broken.

if the mouse moves or a drag overlaps another window, the selection
changes. this happens because of the way selection is happening
inside the selection loop over the children.

my plan here is to move selection to its own method `select_window`
which does the looping and returns the selected window after it
removes it from the children windows. then `on_mouse` can operate
on the selected window if there is one, else does the usual. I'd also
move the offset drag value to be update only if a window is selected.

deselect the window (or drop it) on mouse release.

for rendering or `draw` I'll draw the active window first (if any)
and then the children loop in reverse.